python/03-auth-oauth-server/pyproject.toml
[project]
name = "mcp-tasks-auth"
version = "0.1.0"
description = "MCP server template — Streamable HTTP with Auth0 OAuth (Resource Server pattern)"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
